requirements.txt
rich
requests
pillow
